total refactor, make Move its own file
