+++ src/game/Board.java
@@ -157,53 +157,4 @@ public class Board
         }
         System.out.println();
     }
-
-    public class Move
-    {
-        private GamePiece piece, space;
-        private int fromX, fromY;
-        private int toX, toY;
-
-        public Move(GamePiece piece, int toX, int toY)
-        {
-            this.piece = piece;
-            this.fromX = piece.getCol();
-            this.fromY = piece.getRow();
-
-            this.space = getSpace(toX, toY);
-            this.toX = toX;
-            this.toY = toY;
-        }
-
-        public int getOriginX()
-        {
-            return fromX;
-        }
-
-        public int getOriginY()
-        {
-            return fromY;
-        }
-
-        public int getDestX()
-        {
-            return toX;
-        }
-
-        public int getDestY()
-        {
-            return toY;
-        }
-
-        public GamePiece getOriginPiece()
-        {
-            return piece;
-        }
-
-        public GamePiece getDestPiece()
-        {
-            return space;
-        }
-    }
-
 }
