+++ src/pieces/King.java
@@ -1,5 +1,6 @@
 package pieces;
 import game.Board;
+import game.Move;
 import java.util.ArrayList;
 
 public class King extends GamePiece
@@ -23,7 +24,7 @@ public class King extends GamePiece
     }
 
     @Override
-    public void updateValidMoves(Board board, ArrayList<Board.Move> moves)
+    public void updateValidMoves(Board board, ArrayList<Move> moves)
     {
         for (int dx = -1; dx <= 1; dx++)
         {
@@ -34,7 +35,7 @@ public class King extends GamePiece
                 boolean spaceIsAdjacent  = dx != 0 || dy != 0;
                 boolean spaceIsFriendly  = board.checkSpace(x, y, super.getColor()) == 1;
                 boolean spaceOutOfBounds = board.coordinateOutOfBounds(x, y);
-                if (spaceIsAdjacent && !spaceIsFriendly && !spaceOutOfBounds) { moves.add(board.new Move(this, x, y)); }
+                if (spaceIsAdjacent && !spaceIsFriendly && !spaceOutOfBounds) { moves.add(new Move(board, this, x, y)); }
             }
         }
     }
