+++ src/pieces/Knight.java
@@ -1,5 +1,6 @@
 package pieces;
 import game.Board;
+import game.Move;
 import java.util.ArrayList;
 
 public class Knight extends GamePiece
@@ -20,7 +21,7 @@ public class Knight extends GamePiece
     }
 
     @Override
-    public void updateValidMoves(Board board, ArrayList<Board.Move> moves)
+    public void updateValidMoves(Board board, ArrayList<Move> moves)
     {
         final int COL = super.getCol();
         final int ROW = super.getRow();
@@ -35,8 +36,8 @@ public class Knight extends GamePiece
                 int[] lShapeYPos = {COL+shortEdgeInc, ROW+longEdgeInc};
                 boolean lShapeXInBounds = !board.coordinateOutOfBounds(lShapeXPos[0], lShapeXPos[1]);
                 boolean lShapeYInBounds = !board.coordinateOutOfBounds(lShapeYPos[0], lShapeYPos[1]);
-                if (lShapeXInBounds) { moves.add(board.new Move(this, lShapeXPos[0], lShapeXPos[1])); }
-                if (lShapeYInBounds) { moves.add(board.new Move(this, lShapeYPos[0], lShapeYPos[1])); }
+                if (lShapeXInBounds) { moves.add(new Move(board, this, lShapeXPos[0], lShapeXPos[1])); }
+                if (lShapeYInBounds) { moves.add(new Move(board, this, lShapeYPos[0], lShapeYPos[1])); }
             }
         }
     }
